Implement Exception IPA-GNN (#22)
In the Exception IPA-GNN, instead of the standard execute, branch, aggregate phases of the IPA-GNN, we add a new phase: execute, *raise*, branch, aggregate.

The "raise" phase is similar to the branch phase, except instead of dividing probability mass between the true and false branches, it divides probability mass between the "exception branch" and not.

If the model predicts 0.7 exception, 0.3 not-exception, then 0.7 of the node's probability mass will flow to a new singleton "exception node" ("raise node" in the code). The remaining 0.3 of the node's probability mass will be split between the true branch and false branch as determined by the branch phase.

The exception node behaves similar to the exit node. It is its own parent, and it does not get executed. Once probability mass reaches it, it persists there indefinitely. If no new probability mass is flowing into the exception node, its representation will remain unchanged.

At the end of the forward pass, the model uses only the instruction pointer values of the exception node and exit node to predict whether an exception was raised. It uses the exception node's hidden state to predict the specific exception that was raised.