HParam sweep on GCP TPUs (#26)
This PR contains:

1. All the code changes need to support hparam sweeps
2. A basic 32-TPU random search sweep for a single model (IPAGNN)
3. A new organization structure for event summaries/metadata/checkpoints
4. We're now logging per-run metadata

The organization structure is:
gs://runtime-error-problems-experiments/experiments///
e.g. gs://runtime-error-problems-experiments/experiments/2021-09-07-debugging-fleet/9/10

A "study" is a way of grouping experiments. An experiment is launched via a single command (e.g. a single sweep) on a laptop or dev box. A "run" is a single train run on a TPU.

Studies should be given meaningful names prefixed with the start-date of the study.
Experiment ids are automatically assigned.
Run ids are currently just integers but we should add an hparam summary to the run ids to easily identify runs in tensorboard.

The metadata being logged is:

- command (the command run on the TPU/train machine)
- hostname (the specific TPU or other machine used)
- short_hostname (easier to say name of the host)
- commit (hash of the current commit)
- branch (the current branch name)
- whoami (user launching the experiment)
- datetime (string representation of time experiment was launched at)