Make sure we're not using spans beyond the num_nodes'th span (#48)
If num_nodes is less than max_num_nodes, we pad the span arrays with 0s. This was used by the span encoder to indicate that token 0 was part of a bunch of phantom spans. This only has an effect if using `use_span_index_encoder`, but when that's True, the token embedding of token 0 was computed incorrectly. This commit fixes that.